emit events during cluster deletion

+++ pkg/clusterdeletion/etcdbackupconfig.go
@@ -24,6 +24,7 @@ import (
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	kuberneteshelper "k8c.io/kubermatic/v2/pkg/kubernetes"
 
+	corev1 "k8s.io/api/core/v1"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -32,20 +33,23 @@ func (d *Deletion) cleanupEtcdBackupConfigs(ctx context.Context, cluster *kuberm
 		return nil
 	}
 
-	// always attempt to cleanup, even if the controllers might be disabled now
-	backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
-	if err := d.seedClient.List(ctx, backupConfigs, ctrlruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
-		return fmt.Errorf("failed to get EtcdBackupConfigs: %w", err)
-	}
-
-	if len(backupConfigs.Items) > 0 {
-		for _, backupConfig := range backupConfigs.Items {
-			if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
-				return fmt.Errorf("failed to delete EtcdBackupConfig %q: %w", backupConfig.Name, err)
-			}
+	if cluster.Status.NamespaceName != "" {
+		// always attempt to cleanup, even if the controllers might be disabled now
+		backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
+		if err := d.seedClient.List(ctx, backupConfigs, ctrlruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
+			return fmt.Errorf("failed to get EtcdBackupConfigs: %w", err)
 		}
 
-		return nil
+		if len(backupConfigs.Items) > 0 {
+			for _, backupConfig := range backupConfigs.Items {
+				if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
+					return fmt.Errorf("failed to delete EtcdBackupConfig %q: %w", backupConfig.Name, err)
+				}
+			}
+
+			d.recorder.Eventf(cluster, corev1.EventTypeNormal, "EtcdBackupConfigCleanup", "There are %d EtcdBackupConfig objects waiting for deletion.", len(backupConfigs.Items))
+			return nil
+		}
 	}
 
 	return kuberneteshelper.TryRemoveFinalizer(ctx, d.seedClient, cluster, apiv1.EtcdBackupConfigCleanupFinalizer)
